Code tidy (function return types).
git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@452 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/src/simulator/simexpression.cc
@@ -1226,12 +1226,12 @@ int CNormalPow::Get_Type()
 
 int CNormalPow::Evaluate()
 {
-	return (int)pow(lexpr->Evaluate(), rexpr->Evaluate());
+	return (int)pow((double)lexpr->Evaluate(), rexpr->Evaluate());
 }
 
 int CNormalPow::Evaluate(int* vars)
 {
-	return (int)pow(lexpr->Evaluate(vars), rexpr->Evaluate(vars));
+	return (int)pow((double)lexpr->Evaluate(vars), rexpr->Evaluate(vars));
 }
 
 bool CNormalPow::One_Result()
